requirements.txt
Django~=3.1.3
djangorestframework~=3.12.1
django-randomslugfield~=0.3.0
gunicorn==19.7.1
whitenoise==3.3.1